+++ exercises/book-store/src/lib.rs
@@ -1,4 +1,4 @@
-pub fn lowest_price(books: &[usize]) -> usize {
+pub fn lowest_price(books: &[u32]) -> u32 {
     unimplemented!(
         "Find the lowest price of the bookbasket with books {:?}",
         books
